requirements.txt
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
pytz
pydantic
